Update README.md
This example uses two Arduinos: one as the transmitter (remote control) and the other as the receiver (on the boat). The transmitter reads input from two buttons for controlling the direction of the boat, and it also reads video data from a camera module. It then sends both control signals and video data wirelessly to the receiver